Finished the Bet Result Modal
